refactor(rentals): change some types

+++ src/main/java/com/example/rental_backend/model/Rental.java
@@ -21,8 +21,8 @@ public class Rental {
     private Long id;
 
     private String name;
-    private BigDecimal surface;
-    private BigDecimal price;
+    private Integer surface;
+    private Double price;
     private String picture;
 
     @Column(length = 2000)
@@ -56,19 +56,19 @@ public void setName(String name) {
     this.name = name;
 }
 
-public BigDecimal getSurface() {
+public Integer getSurface() {
     return surface;
 }
 
-public void setSurface(BigDecimal surface) {
+public void setSurface(Integer surface) {
     this.surface = surface;
 }
 
-public BigDecimal getPrice() {
+public Double getPrice() {
     return price;
 }
 
-public void setPrice(BigDecimal price) {
+public void setPrice(Double price) {
     this.price = price;
 }
